Move all table DDL
