+++ src/query/legacy-planners/src/lib.rs
@@ -54,18 +54,7 @@ mod plan_setting;
 mod plan_sink;
 mod plan_sort;
 mod plan_subqueries_set;
-mod plan_table_alter_cluster_key;
-mod plan_table_describe;
-mod plan_table_drop;
-mod plan_table_drop_cluster_key;
-mod plan_table_exists;
-mod plan_table_optimize;
 mod plan_table_recluster;
-mod plan_table_rename;
-mod plan_table_show_create;
-mod plan_table_truncate;
-mod plan_table_undrop;
-mod plan_use_database;
 mod plan_user_alter;
 mod plan_user_create;
 mod plan_user_drop;
@@ -162,20 +151,7 @@ pub use plan_sink::SinkPlan;
 pub use plan_sink::SINK_SCHEMA;
 pub use plan_sort::SortPlan;
 pub use plan_subqueries_set::SubQueriesSetPlan;
-pub use plan_table_alter_cluster_key::AlterTableClusterKeyPlan;
-pub use plan_table_describe::DescribeTablePlan;
-pub use plan_table_drop::DropTablePlan;
-pub use plan_table_drop_cluster_key::DropTableClusterKeyPlan;
-pub use plan_table_exists::ExistsTablePlan;
-pub use plan_table_optimize::OptimizeTableAction;
-pub use plan_table_optimize::OptimizeTablePlan;
 pub use plan_table_recluster::ReclusterTablePlan;
-pub use plan_table_rename::RenameTableEntity;
-pub use plan_table_rename::RenameTablePlan;
-pub use plan_table_show_create::ShowCreateTablePlan;
-pub use plan_table_truncate::TruncateTablePlan;
-pub use plan_table_undrop::UndropTablePlan;
-pub use plan_use_database::UseDatabasePlan;
 pub use plan_user_alter::AlterUserPlan;
 pub use plan_user_create::CreateUserPlan;
 pub use plan_user_drop::DropUserPlan;
