+++ src/query/planner/src/plans/describe_table.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,39 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::sync::Arc;
-
-use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
-use common_meta_app::schema::DropTableReq;
-use common_meta_app::schema::TableNameIdent;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct DropTablePlan {
-    pub if_exists: bool,
-    pub tenant: String,
+pub struct DescribeTablePlan {
     pub catalog: String,
     pub database: String,
-    /// The table name
+    /// The table name.
     pub table: String,
-    pub all: bool,
+    /// The schema description of the output.
+    pub schema: DataSchemaRef,
 }
 
-impl DropTablePlan {
+impl DescribeTablePlan {
     pub fn schema(&self) -> DataSchemaRef {
-        Arc::new(DataSchema::empty())
-    }
-}
-
-impl From<DropTablePlan> for DropTableReq {
-    fn from(p: DropTablePlan) -> Self {
-        DropTableReq {
-            if_exists: p.if_exists,
-            name_ident: TableNameIdent {
-                tenant: p.tenant,
-                db_name: p.database,
-                table_name: p.table,
-            },
-        }
+        self.schema.clone()
     }
 }
