+++ src/query/planner/src/plans/exists_table.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,23 +11,25 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+use std::sync::Arc;
 
+use common_datavalues::DataField;
+use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
+use common_datavalues::ToDataType;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct ShowCreateTablePlan {
-    /// The catalog name
+pub struct ExistsTablePlan {
     pub catalog: String,
-    /// The database name
     pub database: String,
-    /// The table name
     pub table: String,
-    /// The table schema
-    pub schema: DataSchemaRef,
 }
 
-impl ShowCreateTablePlan {
+impl ExistsTablePlan {
     pub fn schema(&self) -> DataSchemaRef {
-        self.schema.clone()
+        Arc::new(DataSchema::new(vec![DataField::new(
+            "result",
+            u8::to_data_type(),
+        )]))
     }
 }
