+++ src/query/planner/src/plans/rename_table.rs
@@ -11,21 +11,28 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 use std::sync::Arc;
 
 use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct DropTableClusterKeyPlan {
+pub struct RenameTablePlan {
     pub tenant: String,
+    pub entities: Vec<RenameTableEntity>,
+}
+
+#[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct RenameTableEntity {
+    pub if_exists: bool,
     pub catalog: String,
     pub database: String,
     pub table: String,
+    pub new_database: String,
+    pub new_table: String,
 }
 
-impl DropTableClusterKeyPlan {
+impl RenameTablePlan {
     pub fn schema(&self) -> DataSchemaRef {
         Arc::new(DataSchema::empty())
     }
