+++ src/query/planner/src/plans/show_create_table.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,23 +11,22 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-use std::sync::Arc;
-
-use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct TruncateTablePlan {
+pub struct ShowCreateTablePlan {
+    /// The catalog name
     pub catalog: String,
+    /// The database name
     pub database: String,
     /// The table name
     pub table: String,
-    pub purge: bool,
+    /// The table schema
+    pub schema: DataSchemaRef,
 }
 
-impl TruncateTablePlan {
+impl ShowCreateTablePlan {
     pub fn schema(&self) -> DataSchemaRef {
-        Arc::new(DataSchema::empty())
+        self.schema.clone()
     }
 }
