+++ src/query/planner/src/plans/truncate_table.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,37 +11,22 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 use std::sync::Arc;
 
 use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
-use common_meta_app::schema::TableNameIdent;
-use common_meta_app::schema::UndropTableReq;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct UndropTablePlan {
-    pub tenant: String,
+pub struct TruncateTablePlan {
     pub catalog: String,
     pub database: String,
+    /// The table name
     pub table: String,
+    pub purge: bool,
 }
 
-impl UndropTablePlan {
+impl TruncateTablePlan {
     pub fn schema(&self) -> DataSchemaRef {
         Arc::new(DataSchema::empty())
-    }
-}
-
-/// The table name
-impl From<UndropTablePlan> for UndropTableReq {
-    fn from(p: UndropTablePlan) -> Self {
-        UndropTableReq {
-            name_ident: TableNameIdent {
-                tenant: p.tenant,
-                db_name: p.database,
-                table_name: p.table,
-            },
-        }
     }
 }
