+++ src/query/service/src/interpreters/interpreter_table_rename.rs
@@ -15,9 +15,9 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_legacy_planners::RenameTablePlan;
 use common_meta_app::schema::RenameTableReq;
 use common_meta_app::schema::TableNameIdent;
+use common_planner::plans::RenameTablePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
