+++ src/query/service/src/interpreters/interpreter_table_undrop.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_legacy_planners::UndropTablePlan;
+use common_planner::plans::UndropTablePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
