+++ src/query/service/src/sql/planner/binder/ddl/table.rs
@@ -33,9 +33,20 @@ use common_datavalues::TypeFactory;
 use common_datavalues::Vu8;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_legacy_planners::OptimizeTableAction;
 use common_legacy_planners::*;
 use common_meta_app::schema::TableMeta;
+use common_planner::plans::AlterTableClusterKeyPlan;
+use common_planner::plans::DescribeTablePlan;
+use common_planner::plans::DropTableClusterKeyPlan;
+use common_planner::plans::DropTablePlan;
+use common_planner::plans::ExistsTablePlan;
+use common_planner::plans::OptimizeTableAction;
+use common_planner::plans::OptimizeTablePlan;
+use common_planner::plans::RenameTableEntity;
+use common_planner::plans::RenameTablePlan;
+use common_planner::plans::ShowCreateTablePlan;
+use common_planner::plans::TruncateTablePlan;
+use common_planner::plans::UndropTablePlan;
 use tracing::debug;
 
 use crate::sql::binder::scalar::ScalarBinder;
