+++ src/query/service/src/sql/planner/binder/mod.rs
@@ -30,12 +30,12 @@ use common_legacy_planners::CreateUserUDFPlan;
 use common_legacy_planners::DropUserPlan;
 use common_legacy_planners::DropUserStagePlan;
 use common_legacy_planners::DropUserUDFPlan;
-use common_legacy_planners::UseDatabasePlan;
 use common_meta_types::UserDefinedFunction;
 use common_planner::plans::CallPlan;
 use common_planner::plans::CreateRolePlan;
 use common_planner::plans::DropRolePlan;
 use common_planner::plans::ShowGrantsPlan;
+use common_planner::plans::UseDatabasePlan;
 use common_planner::MetadataRef;
 pub use scalar::ScalarBinder;
 pub use scalar_common::*;
