+++ src/query/service/src/storages/stage/stage_table.rs
@@ -26,7 +26,6 @@ use common_legacy_planners::Partitions;
 use common_legacy_planners::ReadDataSourcePlan;
 use common_legacy_planners::StageTableInfo;
 use common_legacy_planners::Statistics;
-use common_legacy_planners::TruncateTablePlan;
 use common_meta_app::schema::TableInfo;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::SinkPipeBuilder;
@@ -201,11 +200,7 @@ impl Table for StageTable {
     }
 
     // Truncate the stage file.
-    async fn truncate(
-        &self,
-        _ctx: Arc<dyn TableContext>,
-        _truncate_plan: TruncateTablePlan,
-    ) -> Result<()> {
+    async fn truncate(&self, _ctx: Arc<dyn TableContext>, _: &str, _: bool) -> Result<()> {
         Err(ErrorCode::UnImplement(
             "S3 external table truncate() unimplemented yet!",
         ))
