+++ src/query/service/tests/it/storages/fuse/table.rs
@@ -16,13 +16,12 @@ use std::default::Default;
 
 use common_base::base::tokio;
 use common_exception::Result;
-use common_legacy_planners::AlterTableClusterKeyPlan;
-use common_legacy_planners::DropTableClusterKeyPlan;
 use common_legacy_planners::ReadDataSourcePlan;
 use common_legacy_planners::SourceInfo;
-use common_legacy_planners::TruncateTablePlan;
 use common_meta_app::schema::TableInfo;
 use common_meta_app::schema::TableMeta;
+use common_planner::plans::AlterTableClusterKeyPlan;
+use common_planner::plans::DropTableClusterKeyPlan;
 use databend_query::interpreters::AlterTableClusterKeyInterpreter;
 use databend_query::interpreters::CreateTableInterpreterV2;
 use databend_query::interpreters::DropTableClusterKeyInterpreter;
@@ -178,16 +177,12 @@ async fn test_fuse_table_truncate() -> Result<()> {
     interpreter.execute(ctx.clone()).await?;
 
     let table = fixture.latest_default_table().await?;
-    let truncate_plan = TruncateTablePlan {
-        catalog: fixture.default_catalog_name(),
-        database: fixture.default_db_name(),
-        table: fixture.default_table_name(),
-        purge: false,
-    };
 
     // 1. truncate empty table
     let prev_version = table.get_table_info().ident.seq;
-    let r = table.truncate(ctx.clone(), truncate_plan.clone()).await;
+    let r = table
+        .truncate(ctx.clone(), &fixture.default_catalog_name(), false)
+        .await;
     let table = fixture.latest_default_table().await?;
     // no side effects
     assert_eq!(prev_version, table.get_table_info().ident.seq);
@@ -219,7 +214,9 @@ async fn test_fuse_table_truncate() -> Result<()> {
     assert_eq!(stats.read_rows, (num_blocks * rows_per_block) as usize);
 
     // truncate
-    let r = table.truncate(ctx.clone(), truncate_plan).await;
+    let r = table
+        .truncate(ctx.clone(), &fixture.default_catalog_name(), false)
+        .await;
     assert!(r.is_ok());
 
     // get the latest tbl
