+++ src/query/storages/preludes/src/system/query_log_table.rs
@@ -23,7 +23,6 @@ use common_legacy_planners::Extras;
 use common_legacy_planners::Partitions;
 use common_legacy_planners::ReadDataSourcePlan;
 use common_legacy_planners::Statistics;
-use common_legacy_planners::TruncateTablePlan;
 use common_meta_app::schema::TableIdent;
 use common_meta_app::schema::TableInfo;
 use common_meta_app::schema::TableMeta;
@@ -177,11 +176,7 @@ impl Table for QueryLogTable {
         Ok(())
     }
 
-    async fn truncate(
-        &self,
-        _ctx: Arc<dyn TableContext>,
-        _truncate_plan: TruncateTablePlan,
-    ) -> Result<()> {
+    async fn truncate(&self, _ctx: Arc<dyn TableContext>, _: &str, _: bool) -> Result<()> {
         let mut data = self.data.write();
         *data = VecDeque::new();
         Ok(())
